+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## Unreleased
+- Adds `deepEgal` for comparing plain objects and arrays recursively.  
+  It's still type-safe, so value objects and instances of classes nested in
+  plain objects as compared as `egal` — by value if possible (same constructor
+  and has a `valueOf` function), by reference (`===`) otherwise.
+
+  ```javascript
+  var deepEgal = require("egal").deepEgal
+  function Model(name) { this.name = name }
+
+  deepEgal(42, 42) // => true
+  deepEgal({name: "John"}, {name: "John"}) // => true
+  deepEgal({stats: {age: 13}}, {{stats: age: 13}}) // => true
+  deepEgal([1, 2, 3], [1, 2, 3]) // => true
+  deepEgal(new Model("John"), new Model("John")) // => false
+  deepEgal(new Date(2000, 5), new Date(2000, 5)) // => true
+  ```
+
 ## 1.0.0 (May 25, 2015)
 - No longer considers a primitive and boxed object of the same value to be
   equivalent.